Fix hidden overloaded virtual function in BitParameter.h
accessAsInteger function prototype is defined in BaseParameter
abstract class with an uint32_t type. Nevertheless in BitParameter
the overloaded function take an uin64_t. It seems that GCC accepts
wrongly this mistake but clang don't.
This patch corrects the function prototype accordingly to the one
defined in BaseParameter. It allows the project to be compiled
with clang.

+++ parameter/BitParameter.cpp
@@ -87,7 +87,7 @@ bool CBitParameter::accessAsBoolean(bool& bValue, bool bSet, CParameterAccessCon
     }
 
     // Rely on integer access
-    uint64_t uiValue;
+    uint32_t uiValue;
 
     if (bSet) {
 
@@ -108,7 +108,7 @@ bool CBitParameter::accessAsBoolean(bool& bValue, bool bSet, CParameterAccessCon
 }
 
 // Integer Access
-bool CBitParameter::accessAsInteger(uint64_t& uiValue, bool bSet, CParameterAccessContext& parameterAccessContext) const
+bool CBitParameter::accessAsInteger(uint32_t& uiValue, bool bSet, CParameterAccessContext& parameterAccessContext) const
 {
     uint32_t uiOffset = getOffset();
